as it seems we can not replicate from a worker thread, open issue 5421 on oss redis
