+++ src/Makefile
@@ -48,6 +48,7 @@ CC_SOURCES += $(wildcard $(SOURCEDIR)/util/datablock/*.c)
 CC_SOURCES += $(wildcard $(SOURCEDIR)/util/thpool/*.c)
 CC_SOURCES += $(wildcard $(SOURCEDIR)/index/*.c)
 CC_SOURCES += $(wildcard $(SOURCEDIR)/algorithms/*.c)
+CC_SOURCES += $(wildcard $(SOURCEDIR)/commands/*.c)
 CC_SOURCES += $(wildcard $(SOURCEDIR)/GraphBLASExt/*.c)
 
 # Convert all sources to .o files
