+++ src/commands/commands.h
@@ -1,0 +1,15 @@
+/*
+* Copyright 2018-2019 Redis Labs Ltd. and Contributors
+*
+* This file is available under the Apache License, Version 2.0,
+* modified with the Commons Clause restriction.
+*/
+
+//------------------------------------------------------------------------------
+// Module Commands
+//------------------------------------------------------------------------------
+
+#include "cmd_query.h"
+#include "cmd_delete.h"
+#include "cmd_explain.h"
+#include "cmd_bulk_insert.h"
